Add filemagic integration
... honestly, I'm not a fan of adding an external dependency on libmagic.

Magic is simple enough that a replacement could probably be made in pure Rust,
which I would prefer. But it's a huge clusterf* also, which is probably why
nobody has successfully created an alternative yet.

Might be a neat side project to get started on.

---

This integration just uses the first 1024 bytes (or less if the file is smaller
than that) to determine the file type. If the magic isn't detected within those
first bytes, or needs to read to the end, for example, I don't know what will
happen. Probably just won't be able to guess the file type. Not a big deal.